Update DataReplication
MISMATCH BUG FIX: The bug causing the incorrect matches between providers and the practices they are practicing was the lack of mapping of old ID values with new ID values. When Contact and Account object records were copied in Sandbox, they were receiving brand new IDs, and mapping where the relationships are is fail proof only when using the mapping of these old IDs in the new IDs in the new environment. Before I was using the account name, address, etc that had room for exception cases to cause incorrect matches.